Create a dynamic grid from slider input
Placed the grid creating code in its own function (function() createGrid)

Created an addEventListener to the sliderInput.  (Had to change from
.value because the event listener didn't fire in response to the value).
This event calls  createGrid() anytime the slider is changed so as to
update the grid

Created function clearChildren() to be called before createGrid in the
slider event listener.  Clears all children from the grid so there is a
blank slate to create a grid on