Add clear button and fix erase
Added a button that clears the grid instantly, due to
some mistakes being caused by incorrect logic within the
button listeners which caused the board to get cleared anytime
a button was pressed.  Moved the incorrect logic into its own button

I also fixed an issue with the eraise button which would cause
certain block to get stuck and be non-eraseable or fillable.
It had issues to do with how the eraser filled the cell without
alpha values so anything with an alpha value no longer worked

To fix this I simply added an alpha value to the fill color On
while erasing on rainbow mode, and fixed the if logic to handle
multiple colors at a time